refactor(Registrar): rename classes and methods for clarity
- Rename `AutoInjectServiceRegistrations` to `ServiceRegistrar`.
- Adjust unit tests to reflect the new class and method names.
