+++ src/AutoWire.DI/Extensions/ServiceCollectionExtensions.cs
@@ -12,20 +12,20 @@ public static class ServiceCollectionExtensions
 {
     /// <summary>
     /// Registers services in the provided <see cref="IServiceCollection"/> that are decorated with the <see cref="AutoInjectAttribute"/>.
-    /// This method scans the calling assembly for classes marked with <see cref="AutoInjectAttribute"/> and registers them 
+    /// This method scans the calling assembly for classes marked with <see cref="AutoInjectAttribute"/> and registers them
     /// as services in the specified service collection.
     /// </summary>
     /// <param name="services">The <see cref="IServiceCollection"/> to which services will be added.</param>
     /// <returns>The updated <see cref="IServiceCollection"/> containing the registered services.</returns>
     /// <remarks>
-    /// This method uses <see cref="Assembly.GetCallingAssembly"/> to determine the assembly from which the registration 
-    /// method was called. It is important to ensure that services marked with <see cref="AutoInjectAttribute"/> are 
+    /// This method uses <see cref="Assembly.GetCallingAssembly"/> to determine the assembly from which the registration
+    /// method was called. It is important to ensure that services marked with <see cref="AutoInjectAttribute"/> are
     /// defined in that calling assembly; otherwise, no services will be registered.
     /// </remarks>
     public static IServiceCollection RegisterAutoInjectableServices(this IServiceCollection services)
     {
-        var registrationServices = new AutoInjectServiceRegistrations(services);
-        registrationServices.RegisterFromAssembly(Assembly.GetCallingAssembly());
+        var serviceRegistrar = new ServiceRegistrar(services);
+        serviceRegistrar.RegisterFromAssembly(Assembly.GetCallingAssembly());
         return services;
     }
 }
